requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.63
langchain-ollama==0.2.2
langchain-anthropic==0.3.5
langchain-google-genai==2.1.5
anthropic==0.75.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
rich==13.9.4

# LangGraph for agent workflows (compatible with langchain-core 0.3.x)
langgraph>=0.2.0,<0.3.0
langgraph-checkpoint>=2.0.10,<3.0.0
# RAG and Embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==2.3.1
torch==2.9.1

# File operations and utilities
pathspec==0.12.1
tqdm==4.66.1
